% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{exists_vscode}
\alias{exists_vscode}
\title{Check the VS Code Editor}
\usage{
exists_vscode()
}
\value{
logical class vector that indicates if exists editor VS Code.
}
\description{
Check if exists the VS Code Editor.
}
\examples{
# default
# exists_vscode()
}
