% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{path_extension}
\alias{path_extension}
\title{VS Code Extension Path}
\usage{
path_extension(extension)
}
\arguments{
\item{extension}{character class vector that indicates the id of the extension.}
}
\value{
dataframe class object with VS Code editor extension path information.
}
\description{
Get the VS Code editor extension path.
}
\examples{
path_extension("reditorsupport.r")
}
