% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vscode_extension}
\alias{vscode_extension}
\title{VS Code Extension}
\usage{
vscode_extension()
}
\value{
object of the data frame class that gets the vscode extension.
}
\description{
Get the vscode extension.
}
\examples{
vscode_extension()
}
