% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vscode_version}
\alias{vscode_version}
\title{VS Code Version}
\usage{
vscode_version()
}
\value{
character class vector that indicates the version of the VS Code editor.
}
\description{
Get the VS Code editor version.
}
\examples{
# default
# vscode_version()
}
