% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{exists_vscode}
\alias{exists_vscode}
\title{Check the VS Code Editor}
\usage{
exists_vscode()
}
\value{
The logical class vector indicates if VS Code editor exists.
}
\description{
Check if the VS Code editor exists.
}
\examples{
# default
# exists_vscode()
}
