% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{path_extension}
\alias{path_extension}
\title{VS Code Extension Path}
\usage{
path_extension(extension)
}
\arguments{
\item{extension}{character class vector that indicates the extension id.}
}
\value{
character class vector that indicates the vscode extension path.
}
\description{
Get the path of the vscode extension.
}
\examples{
path_extension("reditorsupport.r")
}
