% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vscode_extension}
\alias{vscode_extension}
\title{VS Code Extension}
\usage{
vscode_extension(type = "basic", format = TRUE)
}
\arguments{
\item{type}{character class vector that indicates the type of search to get the extension with options 'basic' and 'info'.}

\item{format}{logical class vector that indicates whether the column should be formatted with 'TRUE' or 'FALSE' option.}
}
\value{
The object of the dataframe class with information from the VS Code editor extension.
}
\description{
Get the extension that was installed in the VS Code editor.
}
\examples{
# default
# vscode_extension()
# vscode_extension(type = "info")
}
