% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vscode_extension}
\alias{vscode_extension}
\title{VS Code Extension}
\usage{
vscode_extension(type = "basic", format = TRUE)
}
\arguments{
\item{type}{character class vector that indicates the extension search type with option 'basic' and 'info'.}

\item{format}{logical class vector that indicates if must be the column formated with option 'TRUE' or 'FALSE'.}
}
\value{
object of the data frame class that gets the vscode extension.
}
\description{
Get the vscode extension.
}
\examples{
# default
# vscode_extension()
# vscode_extension(type = "info")
}
